* commit 'e15b862293b82df32ec3e8138bf1cbd388450990':
  Lootbox.sol and Test

# Conflicts:
#	packages/hardhat/.openzeppelin/unknown-31337.json

+++ packages/hardhat/contracts/DistributionManager.sol
@@ -26,7 +26,8 @@ contract DistributionManager is Initializable {
     function launchCampaignCommonNFT(
         string memory campaignName,
         string memory tokenURI,
-        uint256 duration,   // days
+
+        uint8 duration,   // days
         uint8 appearance,
         uint8 fightingPower,
         uint8 level,
